Extract pagination code into a helper
